pyproject.toml
[project]
name = "kaggle-heart-disease-competition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "ipykernel>=7.1.0",
    "kaggle>=1.8.3",
    "mlflow>=3.9.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
]
